Renamed extractSnippetValues to extract-snippet-values
