+++ components/landing-page/about-me.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 import {MotionInViewWrapper} from "~/components/ui/motion-in-view-wrapper";
 import type {Snippets} from "~/server/api/routers/snippet";
 import {api} from "~/trpc/server";
-import {extractSnippetValues} from "~/utils/extractSnippetValues";
+import {extractSnippetValues} from "~/utils/extract-snippet-values";
 import type {AboutMeSnippetsFormValues} from "~/utils/validations/about-me";
 
 import {PageSection} from "./page-section";
