+++ components/landing-page/header.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import {MotionInViewWrapper} from "~/components/ui/motion-in-view-wrapper";
 import type {Snippets} from "~/server/api/routers/snippet";
-import {extractSnippetValues} from "~/utils/extractSnippetValues";
+import {extractSnippetValues} from "~/utils/extract-snippet-values";
 import type {HeaderSnippetsFormValues} from "~/utils/validations/header";
 
 import {HeaderButtons} from "./header-buttons";
